mac_requirements.txt
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
macholib==1.16.4
ollama==0.6.1
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreML==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-Vision==12.1
pyperclip==1.11.0
setuptools==80.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
